services/tenant-registry/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
asyncpg==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
structlog==24.1.0
